(#10) Split Sector
