+++ Source/Client/Effects/RocketExplodeEffect.cs
@@ -20,7 +20,7 @@ namespace CodeImp.Bloodmasters.Client
 
 		private Sprite sprite;
 		private Animation ani;
-		private Sector sector;
+		private ClientSector sector;
 		private bool disposed;
 
 		#endregion
@@ -36,7 +36,7 @@ namespace CodeImp.Bloodmasters.Client
 			this.renderpass = 2;
 
 			// Determine current sector
-			sector = General.map.GetSubSectorAt(pos.x, pos.y).Sector;
+			sector = (ClientSector)General.map.GetSubSectorAt(pos.x, pos.y).Sector;
 
 			// Spawn the light
 			if(DynamicLight.dynamiclights)
