+++ Source/Client/Map/ClientSector.cs
@@ -1,0 +1,126 @@
+﻿using System.IO;
+
+namespace CodeImp.Bloodmasters.Client;
+
+public class ClientSector : Sector
+{
+    // Visual Sector
+    private VisualSector vissector = null;
+    private int updatetime = 0;
+    private int firstfloorvertex = -1;
+    private int firstceilvertex = -1;
+    private int numfaces = 0;
+    private ISound sound = null;
+    private bool playmovementsound = false;
+
+    public VisualSector VisualSector { get { return vissector; } set { vissector = value; } }
+    public int FirstFloorVertex { get { return firstfloorvertex; } set { firstfloorvertex = value; } }
+    public int FirstCeilVertex { get { return firstceilvertex; } set { firstceilvertex = value; } }
+    public int NumFaces { get { return numfaces; } set { numfaces = value; } }
+    public bool PlayMovementSound { get { return playmovementsound; } set { playmovementsound = value; } }
+
+    public ClientSector(BinaryReader data, int index, Map map) : base(data, index, map)
+    {
+    }
+
+    protected override void UpdateClientSounds()
+    {
+        var playsound = false;
+
+        // Set lightmap update timer
+        updatetime = General.currenttime;
+
+        // No sound playing?
+        if(sound == null) playsound = true;
+        else if((sound.Filename == SOUND_END) || !sound.Playing) playsound = true;
+
+        // Play start sound?
+        if(playsound && playmovementsound)
+        {
+            // Dispose old sound
+            if(sound != null) sound.Dispose();
+
+            // Play start sound
+            sound = DirectSound.GetSound(SOUND_START, true);
+            sound.Position = new Vector2D(bounds.X + bounds.Width / 2f, bounds.Y + bounds.Height / 2f);
+            sound.Play();
+        }
+    }
+
+    protected override void UpdateLightmaps()
+    {
+        if(vissector != null)
+        {
+            // Update sector lightmaps
+            vissector.UpdateLightmap = true;
+            foreach(Sector s in adjsectors) if(s.VisualSector != null) s.VisualSector.UpdateLightmap = true;
+        }
+    }
+
+    protected override void UpdateClientLighting()
+    {
+        // Time to update?
+        if(updatetime <= General.currenttime)
+        {
+            // Update sector lightmaps
+            UpdateLightmaps();
+
+            // Set timer
+            updatetime = General.currenttime + UPDATE_LIGHTMAP_INTERVAL;
+        }
+
+        // Sound finished playing?
+        if((sound != null) && !sound.Playing && playmovementsound)
+        {
+            // Dispose old sound
+            sound.Dispose();
+
+            // Play moving sound
+            sound = DirectSound.GetSound(SOUND_RUN, true);
+            sound.Position = new Vector2D(bounds.X + bounds.Width / 2f, bounds.Y + bounds.Height / 2f);
+            sound.Play(true);
+        }
+
+        // Running in client mode?
+        if(General.map == this.map)
+        {
+            // Go for all actors
+            foreach(Actor a in General.arena.Actors)
+            {
+                // Actor in this sector and on the floor?
+                if((a.HighestSector == this) && (a.IsOnFloor))
+                {
+                    // Drop on to highest sector
+                    a.DropImmediately();
+                }
+            }
+        }
+    }
+
+    protected override void DropPlayers()
+    {
+    }
+
+    protected override void PlaySounds()
+    {
+        // Play stop sound?
+        if(playstopsound && playmovementsound)
+        {
+            // Dispose old sound
+            if(sound != null) sound.Dispose();
+
+            // Play stop sound
+            sound = DirectSound.GetSound(SOUND_END, true);
+            sound.Position = new Vector2D(bounds.X + bounds.Width / 2f, bounds.Y + bounds.Height / 2f);
+            sound.Play();
+        }
+    }
+
+    public override void Dispose()
+    {
+        if(sound != null) sound.Dispose();
+        sound = null;
+
+        base.Dispose();
+    }
+}
