+++ Source/Shared/PhysicsState.cs
@@ -7,23 +7,14 @@
 
 // This controls the physics in the game
 
-using System;
-using System.Drawing;
 using System.Collections;
-using CodeImp;
-
-#if CLIENT
-using CodeImp.Bloodmasters.Client;
-#else
-using CodeImp.Bloodmasters.Server;
-#endif
 
 namespace CodeImp.Bloodmasters
 {
 	public class PhysicsState
 	{
 		#region ================== Members
-		
+
 		// Settings
 		private float radius;			// Object radius
 		private bool stepup;			// Step up small heights?
@@ -33,26 +24,26 @@ namespace CodeImp.Bloodmasters
 		private bool blocking;			// Blocks other objects?
 		private float height;			// Height of this object
 		private bool isplayer;
-		
+
 		// Position
 		public Vector3D pos;
-		
+
 		// Total Velocity
 		public Vector3D vel;
-		
+
 		// Map
 		private Map map;
-		
+
 		// Collision to render
 		public Collision showcol;
-		
+
 		// Resources
 		private bool disposed = false;
-		
+
 		#endregion
-		
+
 		#region ================== Properties
-		
+
 		public float Radius { get { return radius; } set { radius = value; } }
 		public float Diameter { get { return radius * 2f; } set { radius = value * 0.5f; } }
 		public bool StepUp { get { return stepup; } set { stepup = value; } }
@@ -62,22 +53,22 @@ namespace CodeImp.Bloodmasters
 		public bool Blocking { get { return blocking; } set { blocking = value; } }
 		public bool IsPlayer { get { return isplayer; } set { isplayer = value; } }
 		public float Height { get { return height; } set { height = value; } }
-		
+
 		#endregion
-		
+
 		#region ================== Constructor / Destructor
-		
+
 		// Constructor
 		public PhysicsState(Map map)
 		{
 			// Keep reference
 			this.map = map;
-			
+
 			// Default
 			this.blocking = true;
 			this.isplayer = true;
 		}
-		
+
 		// Disposer
 		public void Dispose()
 		{
@@ -87,11 +78,11 @@ namespace CodeImp.Bloodmasters
 			this.disposed = true;
 			GC.SuppressFinalize(this);
 		}
-		
+
 		#endregion
-		
+
 		#region ================== Methods
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions)
@@ -100,7 +91,7 @@ namespace CodeImp.Bloodmasters
 			Sidedef crossline = null;
 			return ApplyVelocity(wallcollisions, false, clientslist, null, out crossline);
 		}
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions, bool clientcollisions,
@@ -110,7 +101,7 @@ namespace CodeImp.Bloodmasters
 			Sidedef crossline = null;
 			return ApplyVelocity(wallcollisions, clientcollisions, clientslist, thisclient, out crossline);
 		}
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions, bool clientcollisions,
@@ -120,7 +111,7 @@ namespace CodeImp.Bloodmasters
 			ArrayList clientslist = new ArrayList(allclients);
 			return ApplyVelocity(wallcollisions, clientcollisions, clientslist, thisclient, out crossline);
 		}
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions, bool clientcollisions,
@@ -130,7 +121,7 @@ namespace CodeImp.Bloodmasters
 			ArrayList clientslist = new ArrayList(allclients);
 			return ApplyVelocity(wallcollisions, clientcollisions, clientslist, thisclient, out crossline, out hitobj);
 		}
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions, bool clientcollisions,
@@ -139,7 +130,7 @@ namespace CodeImp.Bloodmasters
 			Sidedef crossline = null;
 			return ApplyVelocity(wallcollisions, clientcollisions, allclients, thisclient, out crossline);
 		}
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions, bool clientcollisions,
@@ -149,7 +140,7 @@ namespace CodeImp.Bloodmasters
 			object obj = null;
 			return ApplyVelocity(wallcollisions, clientcollisions, allclients, thisclient, out crossline, out obj);
 		}
-		
+
 		// This moves the position with the total velocity.
 		// Returns true when collided with a wall
 		public bool ApplyVelocity(bool wallcollisions, bool clientcollisions,
@@ -159,32 +150,32 @@ namespace CodeImp.Bloodmasters
 			float stepheight;
 			Vector3D tgt, sv, corr;
 			bool collision = false;
-			
+
 			// No crossings yet
 			crossline = null;
 			hitobj = null;
-			
+
 			// Not disposed?
 			if(disposed) return false;
-			
+
 			// For testing the collision detection
 			showcol = null;
-			
+
 			// DEBUG:
 			//General.DisplayAndLog("Position: " + pos.x + ", " + pos.y + "   Velocity: " + vel.x + ", " + vel.y);
-			
+
 			// Copy velocity
 			sv = vel;
-			
+
 			// Determine target coordinates without collision
 			tgt = pos + sv;
-			
+
 			// Determine step up height
 			if(stepup) stepheight = Consts.MAX_STEP_HEIGHT; else stepheight = 0f;
-			
+
 			// Make collisions array
 			ArrayList colls = new ArrayList();
-			
+
 			// Player collision detection?
 			if(clientcollisions)
 			{
@@ -199,7 +190,7 @@ namespace CodeImp.Bloodmasters
 					}
 				}
 			}
-			
+
 			// Wall collision detection?
 			if(wallcollisions)
 			{
@@ -209,80 +200,80 @@ namespace CodeImp.Bloodmasters
 				{
 					// Get the linedef
 					Linedef ld = (Linedef)lines[i];
-					
+
 					// Make possible collision
 					WallCollision wc = new WallCollision(ld, pos, sv, radius, height, stepheight, isplayer);
 					colls.Add(wc);
-					
+
 					// Return the crossing sidedef, if crossing
 					if(!wc.IsColliding && wc.IsCrossing) crossline = wc.CrossSide;
 				}
 			}
-			
+
 			// Sort the collisions by order in which we will collide with them
 			// The collisions we dont actually collide with will be sorted to the end
 			colls.Sort();
-			
+
 			// DEBUG:
 			//string wallslist = "";
 			//foreach(Collision c in colls) if(c is WallCollision) wallslist += (c as WallCollision).Line.Index + "(" + c.Distance + "), ";
 			//General.DisplayAndLog("Sorted walls: " + wallslist);
-			
+
 			// Go for all possible collisions
 			for(int i = 0; i < colls.Count; i++)
 			{
 				// Get the collisions
 				Collision coll = (Collision)colls[i];
-				
+
 				// For testing the collision detection
 				if(showcol == null) showcol = coll;
-				
+
 				// Will we collide with this?
 				if(coll.IsColliding)
 				{
 					// Collision!
 					collision = true;
-					
+
 					// DEBUG:
 					//General.DisplayAndLog("COLLISION! Line: " + (coll as WallCollision).Line.Index);
-					
+
 					// Return collision object
 					hitobj = coll.CollideObj;
-					
+
 					// Apply new position at collision
 					pos.Apply2D(coll.NewObjPos);
-					
+
 					// DEBUG:
 					//General.DisplayAndLog("Correction: " + pos.x + ", " + pos.y);
-					
+
 					// Continue after collision?
 					if(redirect)
 					{
 						// Choose the bounce or slide vector
 						if(bounce) corr = coll.GetBounceVector(); else corr = coll.GetSlideVector();
-						
+
 						// Scale by friction
 						corr.Scale(friction);
-						
+
 						// Apply new velocity at collision
 						sv.Apply2D(corr);
-						
+
 						// Apply change to source velocity as well
 						vel.Apply2D(corr);
-						
+
 						// DEBUG:
 						//General.DisplayAndLog("New velocity: " + vel.x + ", " + vel.y);
-						
+
 						// Go for all following collisions to update
 						for(int k = i + 1; k < colls.Count; k++)
 						{
 							// Get old collision
 							Collision oldcoll = (Collision)colls[k];
-							
+
 							// Update collision
 							colls[k] = oldcoll.Update(pos, sv);
 						}
-						
+
 						// Sort the lines again, except for those already tested
 						if(i < colls.Count - 1) colls.Sort(i + 1, colls.Count - (i + 1), null);
 					}
@@ -291,7 +282,7 @@ namespace CodeImp.Bloodmasters
 						// Zero velocity
 						vel = new Vector3D(0f, 0f, vel.z);
 						sv = new Vector3D(0f, 0f, sv.z);
-						
+
 						// Done
 						break;
 					}
@@ -304,16 +295,16 @@ namespace CodeImp.Bloodmasters
 					#endif
 				}
 			}
-			
+
 			// Apply the remaining velocity
 			pos.x += sv.x;
 			pos.y += sv.y;
 			pos.z += sv.z;
-			
+
 			// Return collision result
 			return collision;
 		}
-		
+
 		#endregion
 	}
 }
